- Home route set for main screen - Navigation title displayed always

+++ week5/src/main/java/com/chichi289/week5/navigation/Screen.kt
@@ -10,6 +10,6 @@ sealed class Screen(val route: String) {
     }
 
     object PostDetail : Screen("post_detail/{$KEY_POST_ID}") {
-        fun postId(postId: Int) = "post_detail/$postId"
+        fun postId(postId: Long) = "post_detail/$postId"
     }
 }
